+++ changelog.md
@@ -4,6 +4,7 @@ Optional Terms:
 updated, created, *fix(todo), *debug(todo), *add(todo)
 
 v1.0 
+GitHub Name: Map---part-1
 Name: RPG - MAP
 added - map - nested lists
 added - global variables row and col
@@ -14,14 +15,29 @@ added - options to walk's sub menu "N,S,E,W" in movement()
 fixed - movement so player can not walk off "map"
 
 v2.0 
-Name: RPG: Castal Adventure
-added - empty list to function as an inventory of objects
+GitHub Name: Map---part-2
+created - empty list to function as an inventory of objects
+created - objects without a dictionary
+added - conditional statements to deal with objects
+
+v3.0 
+GitHub Name: Map---Dictionarys
+Name: RPG: Castal Adventure!
 added - dictionary for tile/map attributes (description strings)
 fixed - room description for main menu
-created - object dictionary
+updated - object data to a dictionary
 added - 'look' as a main menu option
 added - inspectRoom() function to deal with objects in the game
 added - a fuction to deal with the chest object
 added - a function to deal with the key object
 added - docStrings for each of the functions
 
+v4.0
+GitHub Name: Map---Librarys
+Name: RPG: Castal Adventure!
+added - library random so the key can generate in a random room
+added - libary sys so that 'quit' options more efficently end
+        the game
+removed - while loop variables
+added - messages dictionary for repetative messages in the game
+fixed - DocString line ups